requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.0
billiard==3.5.0.2
boto==2.46.1
boto3==1.4.4
botocore==1.5.35
celery==4.0.2
Django==1.10.5
django-storages==1.5.2
djangorestframework==3.5.4
docutils==0.13.1
gunicorn==19.7.1
jmespath==0.9.2
kombu==4.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.1.10
python-dateutil==2.6.0
python-http-client==2.2.1
pytz==2016.10
s3transfer==0.1.10
sendgrid==3.6.5
six==1.10.0
vine==1.1.3
virtualenv==15.1.0
